test/web-platform-tests/to-upstream/html/semantics/embedded-content/the-img-element/remove-attribute-src.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Removing src attribute from img</title>
<link rel="author" title="atsikov" href="mailto:alexey.tsikov@gmail.com">
<link rel=help href="https://html.spec.whatwg.org/#the-img-element">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
"use strict";

async_test(t => {
  const image = new window.Image();
  image.onload = () => {
    image.onload = t.unreached_func("onload should not be called with removed src");
    image.onerror = t.unreached_func("onerror should not be called with removed src");
    image.removeAttribute("src");

    assert_equals(image.width, 0);
    assert_equals(image.height, 0);
    assert_equals(image.complete, false);
    assert_equals(image.src, "");
    assert_equals(image.currentSrc, "");
  };

  image.src = "data:image/gif;base64,R0lGODlhAQABAIAAAMLCwgAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==";

  t.step_timeout(() => {
    // JSDOM specific: in case image loading is not supported (no canvas api), run checks out of onload handler
    image.removeAttribute("src");
    assert_equals(image.src, "");
    assert_equals(image.currentSrc, "");
    // end of JSDOM specific

    t.done();
  }, 1000);
}, "img is cleared when src attribute is removed");
</script>
</body>
